Talent Ops/modalgateway/llm-backend/requirements.txt
# LLM Backend Dependencies (Simplified)
# Uses custom supabase_client.py instead of supabase package

# Server
flask>=3.0.0
flask-cors>=4.0.0

# HTTP
requests>=2.31.0

# Utils
python-dotenv>=1.0.0
